Allow setting variable length.
It may be possible for a SAS dataset to specify a variable length as
larger than the longest string that happens to be in the data.  I
haven't seen this in the wild, but this will allow me to initialize an
empty Series, set the length, and then read data from an XPT file.