Track version in setup.cfg
There was a bit of a circular import problem caused by adding the pandas
dependency.  If pip tries to import the module to read the version info
before it has installed all dependencies, installation fails.  So, I
moved the version info to setup.cfg and in the code look up the info
from the installation.