Simplify format parsing and display.
I don't have good examples of XPORT files with formats and informats.  I
must guess at what the names look like.  Going through the documentation
of informats, it appears that if you drop the "w.d" portion of names,
no name is longer than 8 characters, including the "$" for character
formats.  My assumption is thus that the "$" is included in the name,
but the "w.d" portion is not.