Project overhaul (again). (#34)
Git commit logs from this overhaul. Good, bad, and ugly:

* Revision of the project structure.

Using the PyPA recommended src layout, this demonstrates the new
load/dump API.  One other major change is pasting the test data into the
Python code for easier verification and modification.

* load/loads returns a library, which has members

* Note installation and contribution instructions.

* Note Python version requirement.

* Add data validation tests.

Variable type, name length, and label length.

* Ignore local data

* Show more logs during test

* Embrace Pandas dataframes

Assuming users will want to use Pandas dataframes for all library
members simplifies the parsing logic dramatically, as we no longer need
to worry about rows vs columns and whether someone wants to use
generators.

Another major change is handling multiple members in a library.  This
has a big impact on the API.

* Provide Pandas DataFrame and Series accessors.

Rather than using composition to create ``Member`` classes that contain
Pandas dataframes, it seems cleaner to provide a ``sas`` accessor for
Pandas dataframes and series. Each dataset now behaves exactly like a
regular dataframe, but has the capacity for adding SAS-style metadata.

* Add docstring

* Allow setting variable length.

It may be possible for a SAS dataset to specify a variable length as
larger than the longest string that happens to be in the data.  I
haven't seen this in the wild, but this will allow me to initialize an
empty Series, set the length, and then read data from an XPT file.

* Allow setting variable number and position.

As I work on converting the file parsing code to the new Pandas accessor
implementation, it's nice to be able to store all metadata on the
accessor objects while reading the file.

* Track version in setup.cfg

There was a bit of a circular import problem caused by adding the pandas
dependency.  If pip tries to import the module to read the version info
before it has installed all dependencies, installation fails.  So, I
moved the version info to setup.cfg and in the code look up the info
from the installation.

* Extend Pandas objects with inheritance.

While the Pandas documentation discourages inheritance, it seems to be
the only way to (mostly) retain metadata.  Using the accessor namespace
technique is bogus, since it silently discards the meta instance after
a variety of usage.

* Include bool columns as SAS numeric.

Otherwise `df.info` causes problems, which makes logging difficult.

Also, I fixed spurious warnings occurring with empty datasets.

* Fix namestr pack/unpack.

* Note differences between v5, v8, and v9.

* Simplify format parsing and display.

I don't have good examples of XPORT files with formats and informats.  I
must guess at what the names look like.  Going through the documentation
of informats, it appears that if you drop the "w.d" portion of names,
no name is longer than 8 characters, including the "$" for character
formats.  My assumption is thus that the "$" is included in the name,
but the "w.d" portion is not.

* Fix format comparison test.

This is unsatisfactory, but it'll have to do until I learn more about
SAS formats.  I'm not sure how left- and right-alignment is specified
with format names.

* Fix member header encode to XPT.

* Fix observation encoding.

* Finally! Fix Pandas extensions.

What a mess.  Pandas does not make it easy to attach metadata to Series
and DataFrames.  Fragile fragile fragile.  Lots of awkward code.

* Allow decoding non-ASCII text.

* Add debug logs for metadata, info for parsing.

* Test and fix dtype coercion/validation.

* Test and fix metadata str length validation

* Add datetime validation for created, modified.

* Test and fix CLI

* Test and fix the old interface.

* Fix issue converting numeric column names.

* Update examples for Pandas subclass design.