Fix `Dataset.label` issues.
A `Dataset.label` property provides compatibility with the old
interface.  I doubt many people started using `dataset_label` in the
last few weeks, but to be safe I left both and provided a deprecation
warning.

I added tests for both version 5/6 and version 8/9 to check the
combinations of specifying a data set label, a variable label, both, and
neither.