Enable changing the text encoding
This is a "beta" feature, and the interface is in flux.  It's awkward to
weave an encoding argument through every class and method, especially
with the data and metadata using different text encodings.  So, I chose
the "practicality beats purity" axiom of the Zen of Python and created
two global variables that configure the data and metadata text
encodings, and a context manager to toggle them.